refactor: move `ZeroCopyHeader`, `AccountLoaderExt` trait to `utils`
