+++ programs/restaking/src/modules/common/mod.rs
@@ -1,7 +1,5 @@
 mod ed25519_verification;
 mod token_program;
-mod zero_copy;
 
 pub use ed25519_verification::*;
 pub use token_program::*;
-pub use zero_copy::*;
