+++ programs/restaking/src/utils.rs
@@ -1,7 +1,6 @@
 #![allow(unexpected_cfgs)]
-use anchor_lang::prelude::*;
-
-use crate::errors::ErrorCode;
+use anchor_lang::solana_program;
+use anchor_lang::{prelude::*, ZeroCopy};
 
 pub trait PDASeeds<const N: usize> {
     const SEED: &'static [u8];
@@ -13,6 +12,138 @@ pub trait PDASeeds<const N: usize> {
         let mut signer_seeds = self.seeds().to_vec();
         signer_seeds.push(std::slice::from_ref(self.bump_ref()));
         signer_seeds
+    }
+}
+
+/// Zero-copy account that has header (data-version, bump).
+pub trait ZeroCopyHeader: ZeroCopy {
+    /// Offset of bump (8bit)
+    fn bump_offset() -> usize;
+}
+
+/// An extension trait for [AccountLoader].
+pub trait AccountLoaderExt<'info> {
+    /// Sets zero-copy header when initializing the account
+    /// without enough account data size provided.
+    ///
+    /// This operation is almost equivalent to [load_init](AccountLoader::load_init),
+    /// but skips bytemuck pointer type casting and accesses directly
+    /// to byte array using offset.
+    fn initialize_zero_copy_header(&mut self, bump: u8) -> Result<()>;
+
+    /// Reads bump directly from data without
+    /// borsh deserialization or bytemuck type casting.
+    fn bump(&self) -> Result<u8>;
+
+    /// Realloc account to add extra amount of data size.
+    /// It can only add max 10KB([`MAX_PERMITTED_DATA_INCREASE`]) at once.
+    ///
+    /// [`MAX_PERMITTED_DATA_INCREASE`]: solana_program::entrypoint::MAX_PERMITTED_DATA_INCREASE
+    fn expand_account_size_if_needed(
+        &self,
+        payer: &Signer<'info>,
+        system_program: &Program<'info, System>,
+        desired_account_size: Option<u32>,
+        assert_size: bool,
+    ) -> Result<()>;
+}
+
+impl<'info, T: ZeroCopyHeader + Owner> AccountLoaderExt<'info> for AccountLoader<'info, T> {
+    fn initialize_zero_copy_header(&mut self, bump: u8) -> Result<()> {
+        if !self.as_ref().is_writable {
+            return Err(ErrorCode::AccountNotMutable.into());
+        }
+
+        // The discriminator should be zero, since we're initializing.
+        let mut data = self.as_ref().try_borrow_mut_data()?;
+        let mut disc_bytes = [0u8; 8];
+        disc_bytes.copy_from_slice(&data[..8]);
+        let discriminator = u64::from_le_bytes(disc_bytes);
+        if discriminator != 0 {
+            return Err(ErrorCode::AccountDiscriminatorAlreadySet.into());
+        }
+
+        // Safety check
+        let offset = 8 + T::bump_offset();
+        if data.len() < offset + 1 {
+            return Err(ErrorCode::ConstraintSpace.into());
+        }
+
+        // Sets bump.
+        data[offset] = bump;
+
+        Ok(())
+    }
+
+    fn bump(&self) -> Result<u8> {
+        let data = self.as_ref().try_borrow_data()?;
+
+        // Safety check
+        let offset = 8 + T::bump_offset();
+        if data.len() < offset + 1 {
+            return Err(ErrorCode::ConstraintSpace.into());
+        }
+
+        Ok(data[offset])
+    }
+
+    fn expand_account_size_if_needed(
+        &self,
+        payer: &Signer<'info>,
+        system_program: &Program<'info, System>,
+        desired_account_size: Option<u32>,
+        assert_size: bool,
+    ) -> Result<()> {
+        let account_info = self.as_ref();
+
+        let current_account_size = account_info.data_len();
+        let min_account_size = 8 + std::mem::size_of::<T>();
+        let target_account_size = desired_account_size
+            .map(|size| std::cmp::max(size as usize, min_account_size))
+            .unwrap_or(min_account_size);
+        let required_realloc_size = target_account_size.saturating_sub(current_account_size);
+
+        msg!(
+            "realloc account size: current={}, target={}, required={}",
+            current_account_size,
+            target_account_size,
+            required_realloc_size
+        );
+
+        if required_realloc_size > 0 {
+            let rent = Rent::get()?;
+            let current_lamports = account_info.lamports();
+            let minimum_lamports = rent.minimum_balance(target_account_size);
+            let required_lamports = minimum_lamports.saturating_sub(current_lamports);
+            if required_lamports > 0 {
+                let cpi_context = CpiContext::new(
+                    system_program.to_account_info(),
+                    anchor_lang::system_program::Transfer {
+                        from: payer.to_account_info(),
+                        to: account_info.clone(),
+                    },
+                );
+                anchor_lang::system_program::transfer(cpi_context, required_lamports)?;
+                msg!("realloc account lamports: added={}", required_lamports);
+            }
+
+            let max_increase = solana_program::entrypoint::MAX_PERMITTED_DATA_INCREASE;
+            let increase = std::cmp::min(required_realloc_size, max_increase);
+            let new_account_size = current_account_size + increase;
+            if assert_size && new_account_size < target_account_size {
+                return Err(crate::errors::ErrorCode::AccountUnmetDesiredReallocSizeError)?;
+            }
+
+            account_info.realloc(new_account_size, false)?;
+            msg!(
+                "account reallocated: current={}, target={}, required={}",
+                new_account_size,
+                target_account_size,
+                target_account_size - new_account_size
+            );
+        }
+
+        Ok(())
     }
 }
 
@@ -47,6 +178,6 @@ pub fn timestamp_now() -> Result<i64> {
 /// Truncates null (0x0000) at the end.
 pub fn from_utf8_trim_null(v: &[u8]) -> Result<String> {
     Ok(std::str::from_utf8(v)
-        .map_err(|_| ErrorCode::DecodeInvalidUtf8FormatException)?
+        .map_err(|_| crate::errors::ErrorCode::DecodeInvalidUtf8FormatException)?
         .replace('\0', ""))
 }
